Rename: 폴더 이름 변경(beforeLogin, afterLogin)
- (auth) -> (beforeLogin)
- (services) -> (afterLogin)으로 폴더 이름을 변경하여 각 폴더의 의미를 명확히 함
