+++ src/app/(afterLogin)/create-post/page.tsx
