+++ src/app/(afterLogin)/layout.tsx
