+++ src/app/(afterLogin)/post/[postId]/page.tsx
