+++ src/app/@modal/(.)create-post/page.tsx
@@ -1,4 +1,4 @@
-import CreatePostPage from "@/app/(services)/create-post/page";
+import CreatePostPage from "@/app/(afterLogin)/create-post/page";
 import Modal from "@/components/common/Modal";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
