+++ src/app/@modal/(.)post/[postId]/page.tsx
@@ -1,5 +1,5 @@
 import Modal from "@/components/common/Modal";
-import PostDetailPage from "@/app/(services)/post/[postId]/page";
+import PostDetailPage from "@/app/(afterLogin)/post/[postId]/page";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export default function Page(props: any) {
